requirements.txt
fastapi
kiteconnect
pymysql
python-dotenv
pytz
redis
resend
schedule
sqlalchemy
uvicorn
